Beginning to fix breakage after the big merge.
