+++ finite_difference_helpers.py
@@ -257,13 +257,13 @@ def ijk_carpetx(idx4):
         if f == 0:
             pass
         elif f == 1:
-            result += f"+PointDesc::DI[{i}]"
+            result += f("+PointDesc::DI[{i}]")
         elif f == -1:
-            result += f"-PointDesc::DI[{i}]"
+            result += f("-PointDesc::DI[{i}]")
         elif f < 0:
-            result += f"-{-f}*PointDesc::DI[{i}]"
+            result += f("-{-f}*PointDesc::DI[{i}]")
         else:
-            result += f"+{f}*PointDesc::DI[{i}]"
+            result += f("+{f}*PointDesc::DI[{i}]")
     return result
 
 def ijkl_string(idx4, FDparams):
