+++ here.py
@@ -1,3 +1,4 @@
+from __future__ import print_function
 def here(*args):
     """
     Here is a debugging tool. It works like print(), except it prefixes any
